+++ packages/better_player_ios/ios/better_player_ios/Sources/better_player_ios/BetterPlayer.swift
@@ -719,7 +719,7 @@ private var presentationSizeContext = 0
     /// Disables Picture-in-Picture.
     @objc public func disablePictureInPicture() {
         BetterPlayerApi.log(1, "disablePictureInPicture()")
-        setPictureInPicture(true)
+        setPictureInPicture(false)
         if let layer = playerLayerRef {
             layer.removeFromSuperlayer()
             playerLayerRef = nil